pyproject.toml
[tool.poetry]
name = "pricing-comparison-agents"
version = "0.1.0"
description = "Pricing comparison multi agent tool"
authors = ["manisha gandhi","manishagandhi27@gmail.com"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.14"
langchain-openai = "^0.1.8"
python-dotenv = "^1.0.1"
langchain = "^0.2.1"
langchain-community = "^0.2.1"
langchain-anthropic = "^0.1.15"
langchain-google-genai = "^1.0.5"
langchain-google-firestore = "^0.3.0"
firestore = "^0.0.8"
chromadb = "^0.5.0"
tiktoken = "^0.7.0"
sentence-transformers = "^3.0.0"
bs4 = "^0.0.2"
firecrawl-py = "^0.0.13"
langchainhub = "^0.1.18"
wikipedia = "^1.4.0"
tavily-python = "^0.3.3"
langgraph = ">=0.2.20,<0.3"
streamlit = "^1.41.1"
ipython = "^8.31.0"

[tool.ruff]
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
